+++ packages/algob/sample-project/js/test/sample-test.js
@@ -21,7 +21,9 @@ describe("Sample Test", function () {
 		master = new AccountStore(masterBalance);
 		fundReceiver = new AccountStore(minBalance);
 		runtime = new Runtime([master, fundReceiver]);
-
+		// This is an example to retrieve account from config file
+		// if user wants to use it, uncomment the following section
+		// runtime.loadAccountsFromConfig();
 		lsig = runtime.createLsigAccount(feeCheckProgram);
 		lsig.sign(master.account.sk);
 	});